Revert some deprecations involving Classpath interface
