+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/symbols/internal/asm/ClassStubTest.java
@@ -39,7 +39,7 @@ import net.sourceforge.pmd.util.PmdClasspathConfig;
 class ClassStubTest {
 
     private static @NotNull TypeSystem newTypeSystem() {
-        return TypeSystem.usingClasspath(PmdClasspathConfig.pmdClasspath().open());
+        return TypeSystem.usingClasspath(Classpath.forOpenClasspath(PmdClasspathConfig.pmdClasspath().open()));
     }
 
     // while parsing the annotation type, ClassStub's parseLock.ensureParsed()
